method specs and delete folio

+++ controller/DeleteFolio.java
@@ -4,6 +4,8 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.io.File;
 
+import javax.swing.JOptionPane;
+
 import api.IModel;
 import gui.IFrame;
 
@@ -19,10 +21,14 @@ public class DeleteFolio implements ActionListener{
 
 	@Override
 	public void actionPerformed(ActionEvent arg0) {
-		String folioName = frame.getCurrentTab().getName();
-		model.deleteFolio(folioName);
-		File file = new File("src/folios/" + folioName + ".sur");
-		file.delete();
+		int dialogButton = JOptionPane.YES_NO_OPTION;
+		int dialogResult = JOptionPane.showConfirmDialog(null, "Are you sure you want to delete?", "Warning", dialogButton);
+		if(dialogResult == 0) {
+			String folioName = frame.getCurrentTab().getName();
+			model.deleteFolio(folioName);
+			File file = new File("src/folios/" + folioName + ".sur");
+			file.delete();
+		}
 		
 	}
